Formated Map return method and CSVReader fix
In a version of the program I named 'V2', I made a method which returns a formated Map of the file data, as CSVReader was having issues.

I also fixed the CSVReader issues, as I found out that it had a dependency, which I've added, although I won't be modifying the first version now that I made what I needed in the second.